优化代码

+++ app/src/main/java/statusbar/lyric/activity/page/MainPage.kt
@@ -36,7 +36,7 @@ class MainPage : BasePage() {
     private fun checkApi() {
         ActivityTools.checkInstalled("cn.lyric.getter").isNotNull {
             val value = it.versionName.split(".")
-            val apiVersion = value[value.lastIndex].toInt()
+            val apiVersion = value[value.lastIndex].toLong()
             if (apiVersion != BuildConfig.apiVersion) {
                 TextSA(textId = R.string.NoSupportedVersionLyricGetter, tipsId = R.string.clickToInstall, onClickListener = {
                     ActivityTools.openUrl("https://github.com/xiaowine/Lyric-Getter/")
